pyproject.toml
[project]
name = "thesis-hub-be"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "black>=25.12.0",
    "email-validator>=2.3.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.14.9",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "uuid7>=0.1.0",
    "uvicorn>=0.38.0",
]
